alpha build

+++ src/components/options/options.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import {ShiftProvider} from "../../providers/shift/shift";
-import {ToastController} from "ionic-angular";
+import {AlertController, ToastController} from "ionic-angular";
 
 /**
  * Generated class for the OptionsComponent component.
@@ -16,18 +16,36 @@ export class OptionsComponent {
 
 
   constructor(private shiftProvider: ShiftProvider,
-              private toastCtrl: ToastController,) {
+              private toastCtrl: ToastController,
+              private alertCtrl: AlertController) {
   }
 
   deleteAll() {
-    this.shiftProvider.deleteAll().then(() => {
-      let toast = this.toastCtrl.create({
-        message: 'Shifts deleted!',
-        position: 'top',
-        duration: 3000
-      });
-      toast.present();
-    })
+    let confirm = this.alertCtrl.create({
+      title: 'Really delete all shifts?',
+      message: '<strong>This action cannot be undone!</strong>',
+      buttons: [
+        {
+          text: 'Cancel',
+          handler: () => {
+          }
+        },
+        {
+          text: 'Delete',
+          handler: () => {
+            this.shiftProvider.deleteAll().then(() => {
+              let toast = this.toastCtrl.create({
+                message: 'Shifts deleted!',
+                position: 'top',
+                showCloseButton: true,
+                duration: 3000
+              });
+              toast.present();
+            })
+          }
+        }
+      ]
+    });
+    confirm.present();
   }
-
 }
